RY3T-Module-Automation/requirements.txt
paho-mqtt
requests
simple_pid